models/ai-mentor-chatbot/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-google-genai>=0.0.10
google-generativeai>=0.3.0
pdfplumber>=0.10.0
streamlit>=1.28.0
